chore: moved weather types to own file in weather-easter-egg feat: added custom hook useWeather
