+++ src/components/content/heroes/aurora/aurora-types.ts
@@ -3,11 +3,3 @@ export enum AuroraType {
   Blue = 'blue',
   BrightBlue = 'bright-blue',
 }
-
-export enum WeatherType {
-  Sunny = 'Sunny',
-  Cloudy = 'Cloudy',
-  Rainy = 'Rainy',
-  Snowy = 'Snowy',
-  NotSet = 'Not set',
-}
