+++ src/components/content/heroes/aurora/weather-easter-egg/use-weather.ts
@@ -1,0 +1,8 @@
+import { useState } from 'react';
+import { WeatherType } from './weather-types';
+
+export const useWeather = () => {
+  const [weather, setWeather] = useState<WeatherType>(WeatherType.NotSet);
+
+  return { weather, setWeather };
+};
