+++ src/components/content/heroes/aurora/weather-easter-egg/weather-types.ts
@@ -1,0 +1,7 @@
+export enum WeatherType {
+  Sunny = 'Sunny',
+  Cloudy = 'Cloudy',
+  Rainy = 'Rainy',
+  Snowy = 'Snowy',
+  NotSet = 'Not set',
+}
